requirements.txt
python-telegram-bot==13.7
pandas==1.4.2
requests==2.27.1
python-dotenv==0.20.0